--- src/main/java/com/project/socket/post/service/usecase/PostSaveInfo.java
@@ -1,14 +1,0 @@
-package com.project.socket.post.service.usecase;
-
-import com.project.socket.post.model.PostMeeting;
-import com.project.socket.post.model.PostType;
-
-public record PostSaveInfo(
-        String title,
-        String postContent,
-        PostType postType,
-        PostMeeting postMeeting,
-        Long userId
-
-) {
-}
